Moved core/metadata to internal/core/metadata
