+++ internal/core/metadata/Attribution.txt
