+++ internal/core/metadata/const.go
