+++ internal/core/metadata/interfaces/db.go
