+++ internal/core/metadata/raml/core-metadata.raml
