+++ internal/core/metadata/rest_addressable.go
