+++ internal/core/metadata/rest_command.go
