+++ internal/core/metadata/router.go
