+++ internal/core/metadata/yaml_deviceprofile.go
