+++ internal/pkg/db/test/db_metadata.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
-	"github.com/edgexfoundry/edgex-go/core/metadata/interfaces"
+	"github.com/edgexfoundry/edgex-go/internal/core/metadata/interfaces"
 	"gopkg.in/mgo.v2/bson"
 )
